Add Layout component wrapps CartContainer
