+++ Week-1/src/App.js
@@ -1,35 +1,11 @@
 import React, { Fragment } from 'react'
-import { BrowserRouter as Router, Route, Switch, NavLink } from 'react-router-dom'
-import { mainPath, aboutPath, productsPath } from '~/src/routes/helpers.js'
-import CartContainer from '~/src/containers/CartContainer.js'
-import Basket from '~/src/ProductCard/Basket.js'
+import Layout from '~/src/Layout.js'
+import { BrowserRouter as Router } from 'react-router-dom'
 
-// Views
-import routes from '~/src/routes/index.js'
-
-const routeWithSubroutes = (route, key) =>(
-  <Route {...route} key={key}/>
-)
 const App = () => (
   <Fragment>
       <Router>
-        <CartContainer >
-          <Basket />
-          <ul>
-            <li>
-              <NavLink to={ mainPath() }>Home</NavLink>
-            </li>
-            <li>
-              <NavLink to={ productsPath() }>Products</NavLink>
-            </li>
-            <li>
-              <NavLink to={ aboutPath() }>About</NavLink>
-            </li>
-          </ul>
-          <Switch>
-            { routes.map((route, key) => routeWithSubroutes(route, key))}
-          </Switch>
-        </CartContainer>
+        <Layout />
       </Router>
   </Fragment>
 )
